requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage
Django==2.1.5
django-cors-headers==3.2.1
django-countries==5.5
django_heroku
django-webpack-loader
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
six==1.14.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wrapt==1.11.2
gunicorn==19.7.1
whitenoise==3.3.1
Pillow
